Included argparse.  Still totally unusable.
Trying to set this up so that when you add an option to the parser, it
gets sent to argparse as well.  My part required a variable name and
value already.  Attaching an arg to those brings us pretty close to
what's needed for argparse.

OED just made things harder.  Why did I start with them as an example?
Their query parameter has a prefix and I haven't thought of an elegant
way to deal with that yet.  Adding that in may be a dirty hack.  Maybe
there should be an option to override the query after it's built?