fd() works for directories with spaces.  kinda.
Before this commit spaces in directories didn't work at all.  I changed
IFS and now they kind of work.

You need to specify '\ ' instead of just ' '.  And when you do that it
still doesn't know that the result is a directory, so you have to select
the dir from a list even if it's the only option.  Not ideal.  The first
issue is tolerable but the second needs a fix.