fd() strips slashes.
Tab completed paths were not being found because of the slash.  Stripped
it out before searching.  It's possible this will botch a path with a
legit slash in the name, but making one of those takes some effort to
begin with and if I actually succeed at it, I deserve to have fd fail.

Do issue numbers in description end up in ticket?  #1