Made generic filter function.  Fixed milestones with it.
Fixed #3.  Was trying too hard before.  Decided to get tokens for all issues, and then filter those tokens now that I have strings.  The only exception is the labels array.  Not sure why I didn't do this to begin with.  The only other downside is I have a list of filterable criteria to maintain.  Not sure how else to do this though since some args are filterable and some aren't.  (if you're not getting a single issue, filtering by ticket number is dumb)

Fixed #7.  Milestones get filtered like the rest.