Send filters to github so we don't dl ALL the issues.  #12
First stab at this.  Huge speedup.

Removed regex support.  Most parameters only take one arg.  You can't tell the api to return issues with Jan and Feb milestones.

The exception is labels.  The regex will correctly match all the relevant labels.  But gh's response to a request for those labels is to return issues that have all of them.  This isn't useful.  Not sure if there's going to be a way around this.

Assignee, label, milestone are supported right now.