Made chord_practice not rely on shuf
shuf is not standard.  If it's not available, use sort -R.
sort -R is new.  If it's not available, use wc to get the number of lines, $RANDOM, to get a random line, head -n $RAND, | tail -n 1 to select a line.
Maybe head|tail should have been what we did in the first place.