screenshot-imgur improvements
Refresh seems to be working.

imgur's api is a little inconsistent in how it returns statuses.  Most http errors show up, but success messages don't always work.  Instead of checking for those, check for the value we're actually trying to retrieve.  If it doesn't exist, handle that failure.

Still not sure why this doesn't work on ubuntu.  Runs in a shell, but not cli-board