Moved fmarks to scripts
Wasn't big enough to get its own repo.  But it actually works now.

fmarks.pl parses bookmarks into a dmenu and returns a url
fmarks.sh is a wrapper for calling fmarks.pl that also opens the url in the active browser.
  (failing that, maybe use ps to see what's open?)

Unsure if this will ever do keywords, but at least it works for now.