Began url_compare.sh
Scripts for comparing prod vs dev.  Takes base urls for each plus an output dir as args.  Reads partial urls from stdin.  Composes those into full urls and gets their screenshots.  Also gets a diff mask and average differences between the two.

Next step is to build a small site as a viewing for this.  Should scan all the results and order by the biggest diff.  Show side by side comparisons, with a hover over to show the diff image.

Ideally that site will also be manageable somehow.  ie, a diff could be checked so it wouldn't show up anymore (or sent to bottom) or it could be added to the refresh queue (which would just be the list of partial urls that echo into first script)