Fixed notin-dzen app name
I wanted to be able to use the ap calling notify-send so I could filter off those messages.  Changed the logic to something that actually made some sense.  Now they're fetched consistently and won't break when an app that isn't notify-send does a notification.

I think notin is blocking a little though.  Changing songs in nuvolaplayer is much slower with this running than without.  Also tried a ruby implementation which was slower still.