Cmus uses album art as wallpaper
This is a many-parter.

1. rxvt-background.sh runs in background of an rxvt window.  Reads filenames from a fifo.  When a file comes up, make a copy of it 80% darker.  (Looked stupid at full brightness).  Then sets it to background image via arcane escape code (printf).
2. New cmus theme uses default transparent background.  Setting background colors writes over the wallpaper.
3. Updated cmus_event.  If our urxvt fifo exists (and it may exist after quitting urxvt.  I think rxvt-background.sh needs a trap or something to clean up when killed), appends album art filename, triggering background change.

I was expecting convert to be too slow (took a second or two in a term) but it's a fraction of a second till the change happens in my urxvt window so I'm content to leave it for now.  I'd prefer urxvt's shading option but I never got that to work with a pixmap image.