Try to avoid false positives on album art
Was using levenshtein to get nearest match to album when no matching album was found.  Thought this would help get around painkiller being labeled '(1997)(320kbs) judas priest - painkiller' or something to that effect.  But if the album was missing entirely, taking the next best thing is not a good option.  This commit ignores albums that don't match the input at all.   I'm treating the itunes result as the canonical name and expecting it to be a substring of whatever name the user provides.