rbcal reads/writes to its own bar
Instead of piping, write into a popen'ed bar.  Advantage is that we can
read back in from it instead of piping that bar somewhere else entirely
to handle clicked actions.  The advantage there is that instead of
printing the stuff to display inside the action I can just use the index
of the entry to display.  No more worries about illegal characters!

sleepwalk runs a block for a set number of seconds.  This is where we
read in bar for a while.  Could probably use a better name but nothing
succinct has come to me.

Combined bg and fg into c, which takes a hash.  Not sure if that really
matters, but it cleaned up the conditional around entry.soon?