Let's try dunst!
It's config is way more extensible than expected.  Can match on properties of the notification and either run a script or change the format of the message (or the urgency I suppose but that's less interesting).  Emptying the format will block the message.  So we get all the power of the notin script, but with notification info coming in as a data structure instead of using bash to parse it.

The only thing I don't like so far is that there's no way I can see to edit a notification.  That would probably be done by zeroing the format and running a script that starts a new notify-send.