workspaces script fix
Try to check if a window still exists before reading its properties.
Becuase this is async, it may disappear before we can query it.  This
caues an unrescuable error, because FFI.

Also introduced a delay between an event causing a redraw.  This is
imperceptible and is enough time for x11 to settle down.

Steps to repro the original bug:
  open leafpad
  enter text
  quit
  do not save
1 event closing 2 windows was the pattern causing this problem.