zoom/dk rule debuggery
zoom has three windows.  there's a home screen, the main meeting screen,
and sometimes a splash message that complains about audio.  they do not
always play nicely with my tiling wm of choice, dk.

this setup has been working intermittently, which is more than I can say
for the others.  the floating window is the hardest to target.  the
  meeting window changes titles (or swaps out for another window).  the
  goal is for the main window to tile but to keep the floating window
  ... umm .. floating so it doesn't occupy a window space.

It's not obvious what order dk uses to apply its rules.  Dk is also
behaving inconsistently - rerunning the rc file from a term tends to
work and restarting dk from dkcmd tends not to.  Haven't found a pattern
yet.   Not totally convinced it's dk's fault and not a pattern in my
testing - sometimes consecutive meetings behave differently!

This is incomplete, but still worth checking in because it's progress.